+++ doc/changeLog.md
@@ -1,3 +1,6 @@
+### 2018-04-10
+- DropdownList新增defaultValue
+
 ### 2018-04-09
 
 - 新增ChangeLog